Move existing files into org/pr subdirectory
