+++ peril-settings.json
@@ -5,62 +5,60 @@
         ],
         "disable_github_check": true
     },
-    "rules": {
-        "pull_request, pull_request.labeled, pull_request.unlabeled": "org/common.ts"
-    },
+    "rules": {},
     "repos": {
         "wordpress-mobile/WordPress-iOS": {
             "pull_request, pull_request.labeled, pull_request.unlabeled": [
-                "Automattic/peril-settings@org/ios-macos.ts",
-                "Automattic/peril-settings@org/github/label.ts",
-                "Automattic/peril-settings@org/github/milestone.ts",
-                "Automattic/peril-settings@org/github/diff-size.ts"
+                "Automattic/peril-settings@org/pr/ios-macos.ts",
+                "Automattic/peril-settings@org/pr/github/label.ts",
+                "Automattic/peril-settings@org/pr/github/milestone.ts",
+                "Automattic/peril-settings@org/pr/github/diff-size.ts"
             ]
         },
         "wordpress-mobile/WordPress-Android": {
             "pull_request, pull_request.labeled, pull_request.unlabeled": [
-                "Automattic/peril-settings@org/android.ts",
-                "Automattic/peril-settings@org/github/label.ts",
-                "Automattic/peril-settings@org/github/milestone.ts",
-                "Automattic/peril-settings@org/github/diff-size.ts",
-                "Automattic/peril-settings@org/github/check-login-subtree.ts"
+                "Automattic/peril-settings@org/pr/android.ts",
+                "Automattic/peril-settings@org/pr/github/label.ts",
+                "Automattic/peril-settings@org/pr/github/milestone.ts",
+                "Automattic/peril-settings@org/pr/github/diff-size.ts",
+                "Automattic/peril-settings@org/pr/github/check-login-subtree.ts"
             ]
         },
         "woocommerce/woocommerce-ios": {
             "pull_request, pull_request.labeled, pull_request.unlabeled": [
-                "Automattic/peril-settings@org/ios-macos.ts",
-                "Automattic/peril-settings@org/github/label.ts",
-                "Automattic/peril-settings@org/github/milestone.ts",
-                "Automattic/peril-settings@org/github/diff-size.ts"
+                "Automattic/peril-settings@org/pr/ios-macos.ts",
+                "Automattic/peril-settings@org/pr/github/label.ts",
+                "Automattic/peril-settings@org/pr/github/milestone.ts",
+                "Automattic/peril-settings@org/pr/github/diff-size.ts"
             ]
         },
         "woocommerce/woocommerce-android": {
             "pull_request, pull_request.labeled, pull_request.unlabeled": [
-                "Automattic/peril-settings@org/android.ts",
-                "Automattic/peril-settings@org/github/label.ts",
-                "Automattic/peril-settings@org/github/milestone.ts",
-                "Automattic/peril-settings@org/github/diff-size.ts",
-                "Automattic/peril-settings@org/github/check-login-subtree.ts"
+                "Automattic/peril-settings@org/pr/android.ts",
+                "Automattic/peril-settings@org/pr/github/label.ts",
+                "Automattic/peril-settings@org/pr/github/milestone.ts",
+                "Automattic/peril-settings@org/pr/github/diff-size.ts",
+                "Automattic/peril-settings@org/pr/github/check-login-subtree.ts"
             ]
         },
         "Automattic/simplenote-ios": {
             "pull_request, pull_request.labeled, pull_request.unlabeled": [
-                "Automattic/peril-settings@org/ios-macos.ts",
-                "Automattic/peril-settings@org/github/label.ts",
-                "Automattic/peril-settings@org/github/milestone.ts",
-                "Automattic/peril-settings@org/github/diff-size.ts"
+                "Automattic/peril-settings@org/pr/ios-macos.ts",
+                "Automattic/peril-settings@org/pr/github/label.ts",
+                "Automattic/peril-settings@org/pr/github/milestone.ts",
+                "Automattic/peril-settings@org/pr/github/diff-size.ts"
             ]
         },
         "Automattic/simplenote-macos": {
             "pull_request, pull_request.labeled, pull_request.unlabeled": [
-                "Automattic/peril-settings@org/ios-macos.ts",
-                "Automattic/peril-settings@org/github/label.ts"
+                "Automattic/peril-settings@org/pr/ios-macos.ts",
+                "Automattic/peril-settings@org/pr/github/label.ts"
             ]
         },
         "Automattic/simplenote-android": {
             "pull_request, pull_request.labeled, pull_request.unlabeled": [
-                "Automattic/peril-settings@org/android.ts",
-                "Automattic/peril-settings@org/github/label.ts"
+                "Automattic/peril-settings@org/pr/android.ts",
+                "Automattic/peril-settings@org/pr/github/label.ts"
             ]
         }
     }
